visioncave/vlm/video_understanding.py
from typing import List, Dict, Any, Optional
import torch
from transformers import AutoProcessor, AutoModel
import numpy as np
from PIL import Image

class VideoUnderstanding:
    """Vision-Language Model for video understanding and querying"""
    
    def __init__(
        self,
        model_name: str = "openai/clip-vit-large-patch14",
        device: str = "cuda" if torch.cuda.is_available() else "cpu"
    ):
        """
        Initialize the VLM model
        
        Args:
            model_name: Name/path of the vision-language model
            device: Device to run inference on
        """
        self.device = device
        self.processor = AutoProcessor.from_pretrained(model_name)
        self.model = AutoModel.from_pretrained(model_name).to(device)
        
    def encode_frames(self, frames: List[np.ndarray]) -> torch.Tensor:
        """
        Encode video frames into embeddings
        
        Args:
            frames: List of frames as numpy arrays
            
        Returns:
            Tensor of frame embeddings
        """
        # Convert frames to PIL images
        pil_frames = [Image.fromarray(frame[..., ::-1]) for frame in frames]
        
        # Process images
        inputs = self.processor(images=pil_frames, return_tensors="pt", padding=True)
        inputs = {k: v.to(self.device) for k, v in inputs.items()}
        
        # Get embeddings
        with torch.no_grad():
            outputs = self.model.get_image_features(**inputs)
            
        return outputs
        
    def encode_text(self, query: str) -> torch.Tensor:
        """
        Encode text query into embedding
        
        Args:
            query: Natural language query
            
        Returns:
            Query embedding tensor
        """
        inputs = self.processor(text=[query], return_tensors="pt", padding=True)
        inputs = {k: v.to(self.device) for k, v in inputs.items()}
        
        with torch.no_grad():
            outputs = self.model.get_text_features(**inputs)
            
        return outputs
        
    def search_video(
        self,
        query: str,
        frame_embeddings: torch.Tensor,
        timestamps: List[float],
        top_k: int = 5
    ) -> List[Dict[str, Any]]:
        """
        Search video frames using natural language query
        
        Args:
            query: Natural language query
            frame_embeddings: Pre-computed frame embeddings
            timestamps: List of frame timestamps
            top_k: Number of top results to return
            
        Returns:
            List of matches with timestamps and similarity scores
        """
        # Encode query
        query_embedding = self.encode_text(query)
        
        # Compute similarities
        similarities = torch.nn.functional.cosine_similarity(
            query_embedding.unsqueeze(1),
            frame_embeddings.unsqueeze(0),
            dim=2
        )
        
        # Get top matches
        top_scores, top_indices = similarities[0].topk(top_k)
        
        # Format results
        results = []
        for score, idx in zip(top_scores, top_indices):
            results.append({
                "timestamp": timestamps[idx],
                "score": float(score),
            })
            
        return results
        
    def describe_scene(self, frame: np.ndarray) -> str:
        """
        Generate natural language description of a video frame
        
        Args:
            frame: Input frame as numpy array
            
        Returns:
            Scene description string
        """
        # TODO: Implement scene description using a more suitable model
        # This is a placeholder - we should use a dedicated image captioning model
        return "Scene description not implemented yet"
